requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
python-dotenv
scikit-learn
numpy
google-generativeai
flask-limiter
waitress
